LayoutTests/fast/forms/menulist-submit-without-selection.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description('A unselected option was submitted as fallback. This behavior was removed by the change of webkit.org/b/35056.');

window.jsTestIsAsync = true;

// waitUntilDone() must be called synchronously here: this load submits the form
// (navigating away) before DOMContentLoaded fires, so js-test.js can't arm it in time.
if (window.testRunner)
    testRunner.waitUntilDone();

var parent = document.createElement('div');
document.body.appendChild(parent);
parent.innerHTML = '<form action="">'
    + '<input type=hidden name="submitted" value="true">'
    + '<select name="select">'
    + '<option disabled>Disabled</option>'
    + '</select>'
    + '</form>';

var query = window.location.search;
if (query.indexOf('submitted=true') == -1) {
    var select = document.getElementsByTagName('select')[0];
    select.selectedIndex = 0;
    document.forms[0].submit();
} else {
    shouldBe('query.indexOf("select=Disabled")', '-1');
    finishJSTest();
}
</script>
</body>
</html>
